todolist/requirements.txt
alembic==1.8.1
appdirs==1.4.4
appnope==0.1.3
asgiref==3.5.2
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
bcrypt==4.0.1
bokeh==3.0.1
boto3==1.26.6
botocore==1.29.6
brotlipy==0.7.0
cachelib==0.9.0
certifi==2021.10.8
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.3
cmdstanpy==1.0.8
colorlover==0.3.0
contourpy==1.0.6
convertdate==2.4.0
coverage==6.5.0
crispy-bootstrap5==0.7
cryptography==3.4.7
cufflinks==0.17.3
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
Deprecated==1.2.13
distlib==0.3.2
Django==4.1.4
django-crispy-forms==1.14.0
django-filter==22.1
dnspython==2.2.1
email-validator==1.3.0
entrypoints==0.4
ephem==4.1.3
exceptiongroup==1.0.1
executing==1.0.0
filelock==3.0.12
Flask-WTF==1.0.1
fonttools==4.37.1
geographiclib==1.52
geopy==2.2.0
greenlet==2.0.1
gunicorn==20.1.0
hijri-converter==2.2.4
holidays==0.16
idna==3.4
importlib-metadata==5.0.0
iniconfig==1.1.1
ipykernel==6.15.2
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
jupyter_client==7.3.5
jupyter-core==4.11.1
jupyterlab-widgets==3.0.3
kiwisolver==1.4.4
korean-lunar-calendar==0.3.1
LunarCalendar==0.0.9
lxml==4.9.1
Mako==1.2.3
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
midtransclient==1.4.1
multitasking==0.0.11
nest-asyncio==1.5.5
numerize==0.12
numpy==1.23.4
packaging==21.3
pandas==1.5.1
pandas-datareader==0.10.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pip==22.3.1
plotly==5.10.0
pluggy==1.0.0
prompt-toolkit==3.0.31
prophet==1.1.1
psutil==5.9.2
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycosat==0.6.3
pycparser==2.20
Pygments==2.13.0
PyMeeus==0.5.11
pyOpenSSL==20.0.1
pyotp==2.6.0
pyparsing==2.4.7
PySocks==1.7.1
pytest==7.2.0
pytest-cov==2.12.1
python-dateutil==2.8.1
python-editor==1.0.4
pytotp==3
pytz==2021.3
PyYAML==5.4.1
pyzmq==23.2.1
redis==4.3.4
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
ruamel-yaml-conda==0.15.100
s3transfer==0.6.0
scikit-learn==1.1.2
scipy==1.9.1
setuptools==65.5.0
setuptools-git==1.2
six==1.16.0
sklearn==0.0
SQLAlchemy==1.4.43
sqlparse==0.4.3
stack-data==0.5.0
tenacity==8.0.1
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.1
tqdm==4.61.2
traitlets==5.0.5
typing-extensions==3.10.0.0
urllib3==1.26.7
virtualenv==20.4.7
wcwidth==0.2.5
Werkzeug==2.2.2
wheel==0.37.1
widgetsnbextension==4.0.3
wrapt==1.13.3
xendit==0.1.3
xyzservices==2022.9.0
yfinance==0.1.74
zipp==3.10.0
